# [opt](profile) optimize timer in profile