--- a/fe/fe-core/src/main/java/org/apache/doris/common/profile/SummaryProfile.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/common/profile/SummaryProfile.java
@@ -75,10 +75,6 @@ public class SummaryProfile {
     public static final String EXECUTED_BY_FRONTEND = "Executed By Frontend";
     // Execution Summary
     public static final String EXECUTION_SUMMARY_PROFILE_NAME = "Execution Summary";
-    public static final String ANALYSIS_TIME = "Analysis Time";
-    public static final String JOIN_REORDER_TIME = "JoinReorder Time";
-    public static final String CREATE_SINGLE_NODE_TIME = "CreateSingleNode Time";
-    public static final String QUERY_DISTRIBUTED_TIME = "QueryDistributed Time";
     public static final String INIT_SCAN_NODE_TIME = "Init Scan Node Time";
     public static final String FINALIZE_SCAN_NODE_TIME = "Finalize Scan Node Time";
     public static final String GET_SPLITS_TIME = "Get Splits Time";
@@ -94,6 +90,7 @@ public class SummaryProfile {
     public static final String WAIT_FETCH_RESULT_TIME = "Wait and Fetch Result Time";
     public static final String FETCH_RESULT_TIME = "Fetch Result Time";
     public static final String WRITE_RESULT_TIME = "Write Result Time";
+    public static final String GET_META_VERSION_TIME = "Get Meta Version Time";
     public static final String GET_PARTITION_VERSION_TIME = "Get Partition Version Time";
     public static final String GET_PARTITION_VERSION_COUNT = "Get Partition Version Count";
     public static final String GET_PARTITION_VERSION_BY_HAS_DATA_COUNT = "Get Partition Version Count (hasData)";
@@ -109,8 +106,8 @@ public class SummaryProfile {
     public static final String NEREIDS_OPTIMIZE_TIME = "Nereids Optimize Time";
     public static final String NEREIDS_TRANSLATE_TIME = "Nereids Translate Time";
     public static final String NEREIDS_DISTRIBUTE_TIME = "Nereids Distribute Time";
-    public static final String NEREIDS_GARBAGE_COLLECT_TIME = "Nereids GarbageCollect Time";
-    public static final String NEREIDS_BE_FOLD_CONST_TIME = "Nereids BeFoldConst Time";
+    public static final String NEREIDS_GARBAGE_COLLECT_TIME = "Garbage Collect During Plan Time";
+    public static final String NEREIDS_BE_FOLD_CONST_TIME = "Nereids Fold Const By BE Time";
 
     public static final String FRAGMENT_COMPRESSED_SIZE = "Fragment Compressed Size";
     public static final String FRAGMENT_RPC_COUNT = "Fragment RPC Count";
@@ -143,25 +140,25 @@ public class SummaryProfile {
 
     // The display order of execution summary items.
     public static final ImmutableList<String> EXECUTION_SUMMARY_KEYS = ImmutableList.of(
+            WORKLOAD_GROUP,
             PARSE_SQL_TIME,
+            PLAN_TIME,
+            NEREIDS_GARBAGE_COLLECT_TIME,
             NEREIDS_LOCK_TABLE_TIME,
             NEREIDS_ANALYSIS_TIME,
             NEREIDS_REWRITE_TIME,
+            NEREIDS_BE_FOLD_CONST_TIME,
+            NEREIDS_COLLECT_TABLE_PARTITION_TIME,
             NEREIDS_OPTIMIZE_TIME,
             NEREIDS_TRANSLATE_TIME,
-            NEREIDS_DISTRIBUTE_TIME,
-            WORKLOAD_GROUP,
-            ANALYSIS_TIME,
-            PLAN_TIME,
-            JOIN_REORDER_TIME,
-            CREATE_SINGLE_NODE_TIME,
-            QUERY_DISTRIBUTED_TIME,
             INIT_SCAN_NODE_TIME,
             FINALIZE_SCAN_NODE_TIME,
             GET_SPLITS_TIME,
             GET_PARTITIONS_TIME,
             GET_PARTITION_FILES_TIME,
             CREATE_SCAN_RANGE_TIME,
+            NEREIDS_DISTRIBUTE_TIME,
+            GET_META_VERSION_TIME,
             GET_PARTITION_VERSION_TIME,
             GET_PARTITION_VERSION_BY_HAS_DATA_COUNT,
             GET_PARTITION_VERSION_COUNT,
@@ -188,21 +185,25 @@ public class SummaryProfile {
             TRANSACTION_COMMIT_TIME,
             SYSTEM_MESSAGE,
             EXECUTED_BY_FRONTEND,
-            NEREIDS_BE_FOLD_CONST_TIME,
-            NEREIDS_GARBAGE_COLLECT_TIME,
             SPLITS_ASSIGNMENT_WEIGHT
     );
 
     // Ident of each item. Default is 0, which doesn't need to present in this Map.
     // Please set this map for new profile items if they need ident.
-    public static ImmutableMap<String, Integer> EXECUTION_SUMMARY_KEYS_IDENTATION
+    public static ImmutableMap<String, Integer> EXECUTION_SUMMARY_KEYS_INDENTATION
             = ImmutableMap.<String, Integer>builder()
-            .put(JOIN_REORDER_TIME, 1)
-            .put(CREATE_SINGLE_NODE_TIME, 1)
-            .put(QUERY_DISTRIBUTED_TIME, 1)
-            .put(INIT_SCAN_NODE_TIME, 1)
-            .put(FINALIZE_SCAN_NODE_TIME, 1)
-            .put(GET_SPLITS_TIME, 2)
+            .put(NEREIDS_GARBAGE_COLLECT_TIME, 1)
+            .put(NEREIDS_LOCK_TABLE_TIME, 1)
+            .put(NEREIDS_ANALYSIS_TIME, 1)
+            .put(NEREIDS_REWRITE_TIME, 1)
+            .put(NEREIDS_COLLECT_TABLE_PARTITION_TIME, 1)
+            .put(NEREIDS_OPTIMIZE_TIME, 1)
+            .put(NEREIDS_TRANSLATE_TIME, 1)
+            .put(INIT_SCAN_NODE_TIME, 2)
+            .put(FINALIZE_SCAN_NODE_TIME, 2)
+            .put(GET_SPLITS_TIME, 3)
+            .put(NEREIDS_DISTRIBUTE_TIME, 1)
+            .put(NEREIDS_BE_FOLD_CONST_TIME, 2)
             .put(GET_PARTITIONS_TIME, 3)
             .put(GET_PARTITION_FILES_TIME, 3)
             .put(CREATE_SCAN_RANGE_TIME, 2)
@@ -257,18 +258,6 @@ public class SummaryProfile {
     // timestamp of query begin
     @SerializedName(value = "queryBeginTime")
     private long queryBeginTime = -1;
-    // Analysis end time
-    @SerializedName(value = "queryAnalysisFinishTime")
-    private long queryAnalysisFinishTime = -1;
-    // Join reorder end time
-    @SerializedName(value = "queryJoinReorderFinishTime")
-    private long queryJoinReorderFinishTime = -1;
-    // Create single node plan end time
-    @SerializedName(value = "queryCreateSingleNodeFinishTime")
-    private long queryCreateSingleNodeFinishTime = -1;
-    // Create distribute plan end time
-    @SerializedName(value = "queryDistributedFinishTime")
-    private long queryDistributedFinishTime = -1;
     @SerializedName(value = "initScanNodeStartTime")
     private long initScanNodeStartTime = -1;
     @SerializedName(value = "initScanNodeFinishTime")
@@ -447,6 +436,8 @@ public class SummaryProfile {
 
     private void updateExecutionSummaryProfile() {
         executionSummaryProfile.addInfoString(PARSE_SQL_TIME, getPrettyParseSqlTime());
+        executionSummaryProfile.addInfoString(PLAN_TIME,
+                getPrettyTime(queryPlanFinishTime, parseSqlFinishTime, TUnit.TIME_MS));
         executionSummaryProfile.addInfoString(NEREIDS_LOCK_TABLE_TIME, getPrettyNereidsLockTableTime());
         executionSummaryProfile.addInfoString(NEREIDS_ANALYSIS_TIME, getPrettyNereidsAnalysisTime());
         executionSummaryProfile.addInfoString(NEREIDS_REWRITE_TIME, getPrettyNereidsRewriteTime());
@@ -457,16 +448,6 @@ public class SummaryProfile {
         executionSummaryProfile.addInfoString(NEREIDS_DISTRIBUTE_TIME, getPrettyNereidsDistributeTime());
         executionSummaryProfile.addInfoString(NEREIDS_GARBAGE_COLLECT_TIME, getPrettyNereidsGarbageCollectionTime());
         executionSummaryProfile.addInfoString(NEREIDS_BE_FOLD_CONST_TIME, getPrettyNereidsBeFoldConstTime());
-        executionSummaryProfile.addInfoString(ANALYSIS_TIME,
-                getPrettyTime(queryAnalysisFinishTime, queryBeginTime, TUnit.TIME_MS));
-        executionSummaryProfile.addInfoString(PLAN_TIME,
-                getPrettyTime(queryPlanFinishTime, queryAnalysisFinishTime, TUnit.TIME_MS));
-        executionSummaryProfile.addInfoString(JOIN_REORDER_TIME,
-                getPrettyTime(queryJoinReorderFinishTime, queryAnalysisFinishTime, TUnit.TIME_MS));
-        executionSummaryProfile.addInfoString(CREATE_SINGLE_NODE_TIME,
-                getPrettyTime(queryCreateSingleNodeFinishTime, queryJoinReorderFinishTime, TUnit.TIME_MS));
-        executionSummaryProfile.addInfoString(QUERY_DISTRIBUTED_TIME,
-                getPrettyTime(queryDistributedFinishTime, queryCreateSingleNodeFinishTime, TUnit.TIME_MS));
         executionSummaryProfile.addInfoString(INIT_SCAN_NODE_TIME,
                 getPrettyTime(initScanNodeFinishTime, initScanNodeStartTime, TUnit.TIME_MS));
         executionSummaryProfile.addInfoString(FINALIZE_SCAN_NODE_TIME,
@@ -502,6 +483,7 @@ public class SummaryProfile {
         setTransactionSummary();
 
         if (Config.isCloudMode()) {
+            executionSummaryProfile.addInfoString(GET_META_VERSION_TIME, getPrettyGetMetaVersionTime());
             executionSummaryProfile.addInfoString(GET_PARTITION_VERSION_TIME, getPrettyGetPartitionVersionTime());
             executionSummaryProfile.addInfoString(GET_PARTITION_VERSION_COUNT, getPrettyGetPartitionVersionCount());
             executionSummaryProfile.addInfoString(GET_PARTITION_VERSION_BY_HAS_DATA_COUNT,
@@ -586,18 +568,6 @@ public class SummaryProfile {
         this.queryBeginTime = queryBeginTime;
     }
 
-    public void setQueryAnalysisFinishTime() {
-        this.queryAnalysisFinishTime = TimeUtils.getStartTimeMs();
-    }
-
-    public void setQueryJoinReorderFinishTime() {
-        this.queryJoinReorderFinishTime = TimeUtils.getStartTimeMs();
-    }
-
-    public void setCreateSingleNodeFinishTime() {
-        this.queryCreateSingleNodeFinishTime = TimeUtils.getStartTimeMs();
-    }
-
     public void setInitScanNodeStartTime() {
         this.initScanNodeStartTime = TimeUtils.getStartTimeMs();
     }
@@ -632,10 +602,6 @@ public class SummaryProfile {
 
     public void setCreateScanRangeFinishTime() {
         this.createScanRangeFinishTime = TimeUtils.getStartTimeMs();
-    }
-
-    public void setQueryDistributedFinishTime() {
-        this.queryDistributedFinishTime = TimeUtils.getStartTimeMs();
     }
 
     public void setQueryPlanFinishTime() {
@@ -817,7 +783,7 @@ public class SummaryProfile {
     }
 
     public String getPrettyNereidsLockTableTime() {
-        return getPrettyTime(nereidsLockTableFinishTime, parseSqlStartTime, TUnit.TIME_MS);
+        return getPrettyTime(nereidsLockTableFinishTime, parseSqlFinishTime, TUnit.TIME_MS);
     }
 
     public String getPrettyNereidsAnalysisTime() {
@@ -851,6 +817,11 @@ public class SummaryProfile {
 
     public String getPrettyNereidsDistributeTime() {
         return getPrettyTime(nereidsDistributeFinishTime, nereidsTranslateFinishTime, TUnit.TIME_MS);
+    }
+
+    private String getPrettyGetMetaVersionTime() {
+        long getMetaVersionTime = getPartitionVersionTime + getTableVersionTime;
+        return getPrettyTime(getMetaVersionTime, 0, TUnit.TIME_MS);
     }
 
     private String getPrettyGetPartitionVersionTime() {