### What problem does this PR solve?

- remove timer for legacy planner
- optimize nereids' timer's indentation
- create new group for getting meta version

example:
```
Execution  Summary:
      -  Workload  Group:  normal
      -  Parse  SQL  Time:  1ms
      -  Plan  Time:  4ms
          -  Garbage  Collect  During  Plan  Time:  0ms
          -  Nereids  Lock  Table  Time:  1ms
          -  Nereids  Analysis  Time:  1ms
          -  Nereids  Rewrite  Time:  1ms
              -  Nereids  Fold  Const  By  BE  Time:  0ms
          -  Nereids  Collect  Table  Partition  Time:  0ms
          -  Nereids  Optimize  Time:  1ms
          -  Nereids  Translate  Time:  0ms
              -  Init  Scan  Node  Time:  N/A
              -  Finalize  Scan  Node  Time:  N/A
                  -  Get  Splits  Time:  N/A
                  -  Get  Partitions  Time:  N/A
                  -  Get  Partition  Files  Time:  N/A
              -  Create  Scan  Range  Time:  N/A
          -  Nereids  Distribute  Time:  1ms
      -  Get  Meta  Version  Time:  N/A
          -  Get  Partition  Version  Time:  N/A
          -  Get  Partition  Version  Count  (hasData):  N/A
          -  Get  Partition  Version  Count:  N/A
          -  Get  Table  Version  Time:  N/A
          -  Get  Table  Version  Count:  N/A
```

### Release note

None

### Check List (For Author)

- Test <!-- At least one of them must be included. -->
    - [ ] Regression test
    - [ ] Unit Test
    - [ ] Manual test (add detailed scripts or steps below)
    - [x] No need to test or manual test. Explain why:
        - [x] This is a refactor/code format and no logic has been changed.
        - [ ] Previous test can cover this change.
        - [ ] No code files have been changed.
        - [ ] Other reason <!-- Add your reason?  -->

- Behavior changed:
    - [ ] No.
    - [ ] Yes. <!-- Explain the behavior change -->

- Does this need documentation?
    - [ ] No.
    - [ ] Yes. <!-- Add document PR link here. eg: https://github.com/apache/doris-website/pull/1214 -->

### Check List (For Reviewer who merge this PR)

- [ ] Confirm the release note
- [ ] Confirm test cases
- [ ] Confirm document
- [ ] Add branch pick label <!-- Add branch pick label that this PR should merge into -->